\docType{data}
\name{exdata01}
\alias{exdata01}
\title{A simulated example dataset}
\format{A simulated cohort data with binomial outcome (n=40).}
\usage{
data(exdata01)
}
\description{
A simulated cohort data with binomial outcome.
\itemize{
  \item \code{y}: Dichotomous outcome variable.
  \item \code{x1}: Continuous covariate.
  \item \code{x2}: Binary covariate.
  \item \code{x3}: Binary covariate.
  \item \code{x4}: Binary covariate.
}
}
\keyword{datasets}
