\docType{data}
\name{exdata03}
\alias{exdata03}
\title{A simulated example dataset with missing covariates}
\format{A simulated cohort data with binomial outcome (n=1200). Some covariates involve missing data.}
\usage{
data(exdata03)
}
\description{
A simulated cohort data with binomial outcome. Some covariates involve missing data.
\itemize{
  \item \code{y}: Dichotomous outcome variable.
  \item \code{x1}: Continuous covariate.
  \item \code{x2}: Binary covariate.
  \item \code{x3}: Binary covariate.
  \item \code{x4}: Binary covariate.
}
}
\keyword{datasets}
