\docType{package}
\name{rqlm-package}
\alias{rqlm-package}
\title{The 'rqlm' package.}
\description{
Modified Poisson and least-squares regression analyses for binary outcomes have been standard multivariate analysis methods to estimate risk ratio and risk difference in clinical and epidemiological studies. This R package involves an easy-to-handle function to implement these analyses by simple commands. Also, recent studies have shown the ordinary robust variance estimator possibly has serious bias under small or moderate sample size situations for these methods. This package also provides computational tools to calculate accurate confidence intervals (Noma et al., 2024+).
}
\references{
Cheung, Y. B. (2007). A modified least-squares regression approach to the estimation of risk difference. \emph{American Journal of Epidemiology} \strong{166}, 1337-1344.

Noma, H. et al. (2024+). Bootstrap confidence intervals based on quasi-score statistics for the modified Poisson and least-squares regressions to estimate risk ratio and risk difference. Forthcoming.

Zou, G. (2004). A modified poisson regression approach to prospective studies with binary data. \emph{American Journal of Epidemiology} \strong{159}, 702-706.
}
